requirements-ci.txt
pyyaml>=6.0.1,<7
jsonschema>=4.23.0,<5
